selesai 5 juli

+++ lib/model/getkontribusi.dart
@@ -33,7 +33,7 @@ class Welcome {
 }
 
 class Kontribusi {
-    int tukar;
+    double tukar;
 
     Kontribusi({
         required this.tukar,
